Bug # ????? : iOS and Server Mobile "Unbatching"
We're still batching beacons together for a single transmit, but instead of sending summarized records of similar data, we're now sending all of the raw data points.  This allows us to log all "beacons" in the batch as individual beacons in raw S3 log files.

The iOS Beacon classes were reorganized and standardized to the new Beacon Types.  Each Beacon type is a sub-class of the MPBeacon class, and adds its own methods for serialization when called to.

All of the pre-summariziation code (histogram generation, etc) has been removed, since we're now only sending raw data points.

Batches are now gzipped for transport.

The code is now using ObjClean for standardization, though this isn't enforced in the build (yet).  I didn't clean up all of the files -- only files that I touched in a significant way.

A/B test support was added to mPulse as well.

All of the tests were reworked.  The networking request no longer wait for a fixed timeout, but they poll waiting for a beacon with a max timeout.  Tests were added for each new beacon type as well.

Reviewer: Mukul, Andreas, Matt
Document [y/n]: n

+++ Boomerang/Beacons/MPNetworkErrorBeaconGenerator.m
@@ -17,27 +17,29 @@ static MPNetworkErrorBeaconGenerator *sharedObject = nil;
 static int ERROR_REPLAY_INTERVAL = 30; // In seconds
 
 // URLs that generate different network error
-NSString * const MPPAGENOTFOUND_URL = @"http://67.111.67.24:8080/concertoXYZ";
-NSString * const MPCONNECTION_REFUSED_URL = @"http://67.111.67.24:1200/concerto";
-NSString * const MPUNKNOWN_HOST_URL = @"http://bearsbearsbears123.com";
-NSString * const MPCONNECTION_TIMEOUT_URL = @"http://1.2.3.4:8080/concerto";
+NSString *const MPPAGENOTFOUND_URL = @"http://67.111.67.24:8080/concertoXYZ";
+NSString *const MPCONNECTION_REFUSED_URL = @"http://67.111.67.24:1200/concerto";
+NSString *const MPUNKNOWN_HOST_URL = @"http://bearsbearsbears123.com";
+NSString *const MPCONNECTION_TIMEOUT_URL = @"http://1.2.3.4:8080/concerto";
 
 NSArray *urlList;
 
-// Start the Network Error Beacon Generator
-+ (void) startGenerator
+/**
+ * Start the Network Error Beacon Generator
+ */
++(void) startGenerator
 {
   static dispatch_once_t _singletonPredicate;
   dispatch_once(&_singletonPredicate, ^{
     sharedObject = [[super alloc] init];
     
-    urlList  = @[MPPAGENOTFOUND_URL,
-                 MPUNKNOWN_HOST_URL,
-                 MPCONNECTION_TIMEOUT_URL,
-                 MPCONNECTION_REFUSED_URL];
+    urlList = @[MPPAGENOTFOUND_URL,
+                MPUNKNOWN_HOST_URL,
+                MPCONNECTION_TIMEOUT_URL,
+                MPCONNECTION_REFUSED_URL];
 
     // Dispatch a task to generate Network Errors
-      dispatch_async(dispatch_get_global_queue(DISPATCH_QUEUE_PRIORITY_DEFAULT, (unsigned long)NULL), ^{[self generateNetworkErrorBeacons];});
+    dispatch_async(dispatch_get_global_queue(DISPATCH_QUEUE_PRIORITY_DEFAULT, (unsigned long)NULL), ^{[self generateNetworkErrorBeacons];});
   });
 }
 
@@ -45,7 +47,7 @@ NSArray *urlList;
  * Sends network error to the server by randomizing error count
  * This runs at every 30 seconds interval
  */
-+ (void) generateNetworkErrorBeacons
++(void) generateNetworkErrorBeacons
 {
   // Because this Generator thread is started by a Notification receiver, we cannot
   // read MPConfig ivars at the time of kickoff.
@@ -58,8 +60,8 @@ NSArray *urlList;
   }
   
   // Get random index in urlList
-  NSInteger requestURL1 = arc4random_uniform(urlList.count);
-  NSInteger requestURL2 = arc4random_uniform(urlList.count);
+  long requestURL1 = arc4random_uniform((unsigned int)urlList.count);
+  long requestURL2 = arc4random_uniform((unsigned int)urlList.count);
   
   // Send synchronous request using the urls at the index
   [self sendSynchRequest:[urlList objectAtIndex:requestURL1]];
@@ -73,7 +75,7 @@ NSArray *urlList;
 }
 
 // Send synchronous request
-+ (void) sendSynchRequest: (NSString*) urlString
++(void) sendSynchRequest:(NSString *)urlString
 {
   NSURL *url = [NSURL URLWithString:[urlString stringByAddingPercentEscapesUsingEncoding:NSUTF8StringEncoding]];
   NSURLRequest *request = [NSURLRequest requestWithURL:url cachePolicy:NSURLRequestReloadIgnoringCacheData timeoutInterval:30.0];
